fixed integration tests by testing Mine & Chain directly
